A星算法小错误调整 ComplexEC建立

+++ test/org/socialforce/strategy/impl/AStarPathFinderTest.java
@@ -47,7 +47,7 @@ public class AStarPathFinderTest {
 
     @Test
     public void testMap() throws Exception {
-        double[][]map=new double[][]{{0, 0, 0, 0, 0}, {0, 1, 0, 0, 0}, {0, 0, 0, 1, 0},{0, 0, 1, 1, 0}, {0, 0, 0, 0, 0}};
+        int[][]map=new int[][]{{0, 0, 0, 0, 0}, {0, 1, 0, 0, 0}, {0, 0, 0, 1, 0},{0, 0, 1, 1, 0}, {0, 0, 0, 0, 0}};
         aStarPathFinder = new AStarPathFinder(map,  new BaseAgent(agent_shape, new Velocity2D(0,0)) , new Point2D(3,3));
         Path path = aStarPathFinder.plan_for(new Point2D(3,3));
         System.out.println(path.toString(new Point2D(1,1)));
@@ -55,7 +55,7 @@ public class AStarPathFinderTest {
 
     @Test
     public void testSceneAfterMap() throws Exception {
-        double[][]map=new double[][]{{0, 0, 0, 0, 0}, {0, 1, 0, 0, 0}, {0, 0, 0, 1, 0},{0, 0, 1, 1, 0}, {0, 0, 0, 0, 0}};
+        int[][]map=new int[][]{{0, 0, 0, 0, 0}, {0, 1, 0, 0, 0}, {0, 0, 0, 1, 0},{0, 0, 1, 1, 0}, {0, 0, 0, 0, 0}};
         aStarPathFinder = new AStarPathFinder(map,  new BaseAgent(agent_shape, new Velocity2D(0,0)) , new Point2D(3,3));
         aStarPathFinder = new AStarPathFinder(scene, agent_shape);
         Path path = aStarPathFinder.plan_for(goal);
